banking_system/requirements.txt
appdirs==1.4.4
asgiref==3.3.1
attrs==20.3.0
black==20.8b1
click==7.1.2
Django==3.1.7
djangorestframework==3.12.2
et-xmlfile==1.0.1
flake8==3.8.4
flake8-black==0.2.1
jdcal==1.4.1
jsonschema==3.2.0
mccabe==0.6.1
mypy-extensions==0.4.3
numpy==1.20.1
openpyxl==3.0.6
pandas==1.2.3
pathspec==0.8.1
pkg-resources==0.0.0
pycodestyle==2.6.0
pyflakes==2.2.0
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
regex==2020.11.13
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
